rozbehane automaticke vypisy zacaty admin

+++ src/main/java/com/pis/dao/MaerzDao.java
@@ -1,5 +1,5 @@
 package com.pis.dao;
 
 public interface MaerzDao {
-	public int getG23(int mesiac);
+	public double getG23(int mesiac);
 }
